Rename contourShader file to ContourShader
