+++ ts/Series/Contour/ContourShader.ts
